+++ Changes.md
@@ -1,3 +1,13 @@
+1.1.x.x (relative to 1.1.6.0)
+=======
+
+Fixes
+-----
+
+- Cycles :
+  - Fixed handling of OSL shader component connections (#4926).
+  - Fixed crash when attempting to use a non-existent OSL shader.
+
 1.1.6.0 (relative to 1.1.5.0)
 =======